Mod menu!!!

+++ src/Caspian/Render/RenderUtils/Components/SizeComponent.hpp
@@ -6,7 +6,7 @@ class SizeComponent : public Vec2 {
 public:
 	SizeComponent(float x, float y);
 
-	SizeComponent operator*(const int num) {
+	SizeComponent operator*(const float num) {
 		return SizeComponent(this->size.x * num, this->size.y * num);
 	}
 private:
